convert file mode to uint in Linux MirrorProvider
Fixup our use of uint for file permission modes on Linux
in MirrorProvider to match the change in commit
0c84b9323e3d06f51cbfc469d2494abec514d226.

+++ MirrorProvider/MirrorProvider.Linux/LinuxFileSystemVirtualizer.cs
@@ -113,7 +113,7 @@ namespace MirrorProvider.Linux
                         {
                             return Result.EIOError;
                         }
-                        ushort fileMode = (ushort)(stat.st_mode & FilePermissions.ALLPERMS);
+                        uint fileMode = (uint)(stat.st_mode & FilePermissions.ALLPERMS);
 
                         Result result = this.virtualizationInstance.WritePlaceholderFile(
                             childRelativePath,
